requirements.txt

python-telegram-bot==20.3
solana
solders
requests
